Remove experimental directory
