+++ eva/optimizer/rules/rules.py
@@ -40,6 +40,7 @@ from eva.optimizer.rules.rules_base import Promise, Rule, RuleType
 from eva.parser.types import JoinType, ParserOrderBySortType
 from eva.plan_nodes.apply_and_merge_plan import ApplyAndMergePlan
 from eva.plan_nodes.create_mat_view_plan import CreateMaterializedViewPlan
+from eva.plan_nodes.exchange_plan import ExchangePlan
 from eva.plan_nodes.explain_plan import ExplainPlan
 from eva.plan_nodes.hash_join_build_plan import HashJoinBuildPlan
 from eva.plan_nodes.nested_loop_join_plan import NestedLoopJoinPlan
@@ -1257,6 +1258,95 @@ class LogicalVectorIndexScanToPhysical(Rule):
         for child in before.children:
             after.append_child(child)
         yield after
+
+
+class LogicalExchangeToPhysical(Rule):
+    def __init__(self):
+        pattern = Pattern(OperatorType.LOGICALEXCHANGE)
+        pattern.append_child(Pattern(OperatorType.DUMMY))
+        super().__init__(RuleType.LOGICAL_EXCHANGE_TO_PHYSICAL, pattern)
+
+    def promise(self):
+        return Promise.LOGICAL_EXCHANGE_TO_PHYSICAL
+
+    def check(self, grp_id: int, context: OptimizerContext):
+        return True
+
+    def apply(self, before: LogicalExchange, context: OptimizerContext):
+        after = ExchangePlan(before.view)
+        for child in before.children:
+            after.append_child(child)
+        yield after
+
+
+class LogicalApplyAndMergeToRayPhysical(Rule):
+    def __init__(self):
+        pattern = Pattern(OperatorType.LOGICAL_APPLY_AND_MERGE)
+        pattern.append_child(Pattern(OperatorType.DUMMY))
+        super().__init__(RuleType.LOGICAL_APPLY_AND_MERGE_TO_PHYSICAL, pattern)
+
+    def promise(self):
+        return Promise.LOGICAL_APPLY_AND_MERGE_TO_PHYSICAL
+
+    def check(self, grp_id: int, context: OptimizerContext):
+        return True
+
+    def apply(self, before: LogicalApplyAndMerge, context: OptimizerContext):
+        apply_plan = ApplyAndMergePlan(before.func_expr, before.alias, before.do_unnest)
+
+        parallelism = 2 if len(Context().gpus) > 1 else 1
+        ray_parallel_env_conf_dict = [
+            {"CUDA_VISIBLE_DEVICES": str(i)} for i in range(parallelism)
+        ]
+
+        exchange_plan = ExchangePlan(
+            inner_plan=apply_plan,
+            parallelism=parallelism,
+            ray_pull_env_conf_dict={"CUDA_VISIBLE_DEVICES": "0"},
+            ray_parallel_env_conf_dict=ray_parallel_env_conf_dict,
+        )
+        for child in before.children:
+            exchange_plan.append_child(child)
+
+        yield exchange_plan
+
+
+class LogicalProjectToRayPhysical(Rule):
+    def __init__(self):
+        pattern = Pattern(OperatorType.LOGICALPROJECT)
+        pattern.append_child(Pattern(OperatorType.DUMMY))
+        super().__init__(RuleType.LOGICAL_PROJECT_TO_PHYSICAL, pattern)
+
+    def promise(self):
+        return Promise.LOGICAL_PROJECT_TO_PHYSICAL
+
+    def check(self, before: LogicalProject, context: OptimizerContext):
+        return True
+
+    def apply(self, before: LogicalProject, context: OptimizerContext):
+        project_plan = ProjectPlan(before.target_list)
+        # Check whether the projection contains a UDF
+        if before.target_list is None or not any(
+            [isinstance(expr, FunctionExpression) for expr in before.target_list]
+        ):
+            for child in before.children:
+                project_plan.append_child(child)
+            yield project_plan
+        else:
+            parallelism = 2 if len(Context().gpus) > 1 else 1
+            ray_parallel_env_conf_dict = [
+                {"CUDA_VISIBLE_DEVICES": str(i)} for i in range(parallelism)
+            ]
+
+            exchange_plan = ExchangePlan(
+                inner_plan=project_plan,
+                parallelism=parallelism,
+                ray_pull_env_conf_dict={"CUDA_VISIBLE_DEVICES": "0"},
+                ray_parallel_env_conf_dict=ray_parallel_env_conf_dict,
+            )
+            for child in before.children:
+                exchange_plan.append_child(child)
+            yield exchange_plan
 
 
 # IMPLEMENTATION RULES END
